+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.0.21
 - [move camera & always ensure selection](https://github.com/rapushka/deck_scaler_rts/pull/29)
+- [spawn the "level" (just a gray rectangle in fact)](https://github.com/rapushka/deck_scaler_rts/pull/31)
 
 # v0.0.20
 - [automatically change unit's mode to manual, on order-click](https://github.com/rapushka/deck_scaler_rts/pull/19)